update_version: add support for forgeversion
Since Fedora 40, the forge macros have supported a %forgeversion macro
which allows including snapshot information in the Version macro instead
of magically prepending it to the %dist tag using %distprefix as was done
previously.
In order to enable this, a new %forgeversion macro was added.
See the has_forgeversion() docstring for more information about
forgeversion.